Move ConfigurationLogger to common module
