+++ sharding-core/sharding-core-common/src/test/resources/logback-test.xml
@@ -26,7 +26,7 @@
         <appender-ref ref="console"/>
     </logger>
     <logger name="org.apache.shardingsphere.underlying.executor.sql.jdbc.executor.ExecutorExceptionHandler" level="off" />
-    <logger name="org.apache.shardingsphere.sharding.core.log.ConfigurationLogger" level="off" />
+    <logger name="org.apache.shardingsphere.underlying.common.log.ConfigurationLogger" level="off" />
     <logger name="ShardingSphere-metadata" level="off" />
     
     <root>
