+++ sharding-spring/sharding-jdbc-orchestration-spring/sharding-jdbc-orchestration-spring-namespace/src/test/resources/logback-test.xml
@@ -25,7 +25,7 @@
     <logger name="org.apache.shardingsphere" level="warn" additivity="false">
         <appender-ref ref="console"/>
     </logger>
-    <logger name="org.apache.shardingsphere.sharding.core.log.ConfigurationLogger" level="off" />
+    <logger name="org.apache.shardingsphere.underlying.common.log.ConfigurationLogger" level="off" />
     <logger name="ShardingSphere-metadata" level="off" />
     
     <root>
